Finish the implementation of the Relation and the DefaultSource. A few tests still failing.
