+++ src/main/scala/com/sparkfits/ReadFits.scala
@@ -36,10 +36,11 @@ object ReadFits {
 
     // Loop over the two HDU of the test file
     for (hdu <- 1 to 2) {
-      val df = spark.readfits
-        .option("HDU", hdu)                 // Index of the HDU
+      val df = spark.read
+        .format("com.sparkfits")
+        .option("hdu", hdu)                 // Index of the HDU
         .option("verbose", true)            // pretty print
-        .option("recordLength", 1 * 1024)   // 1 KB per record
+        .option("recordlength", 1 * 1024)   // 1 KB per record
         .load(args(0).toString)             // File to load
 
       df.show()
